# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dev Connect API",
    "description": "API for developer-focused social network platform where users can create profiles, showcase projects, and connect with other developers.",
    "version": "1.0.0",
    "contact": {
      "name": "Dev Connect Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "components": {
    "schemas": {
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmas2xbuy0000tf8cp8h37lgl"
          },
          "content": {
            "type": "string",
            "example": "This project looks amazing! The UI design is very clean."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "authorId": {
            "type": "string",
            "example": "cmas2xbuy0000tf8cp8h37lgl"
          },
          "projectId": {
            "type": "string",
            "example": "cmas2xbuy0000tf8cp8h37lgl"
          },
          "author": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "avatarUrl": {
                "type": "string",
                "nullable": true
              }
            }
          }
        },
        "required": [
          "id",
          "content",
          "createdAt",
          "authorId",
          "projectId",
          "author"
        ]
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmas2xbuy0000tf8cp8h37lgl"
          },
          "title": {
            "type": "string",
            "example": "Developer Connect Platform"
          },
          "description": {
            "type": "string",
            "example": "A social network for developers to connect and share projects"
          },
          "repoUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://github.com/username/project-repo"
          },
          "liveUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://project-demo.example.com"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["React", "TypeScript", "Node.js"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "authorId": {
            "type": "string",
            "example": "cmas2xbuy0000tf8cp8h37lgl"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Like"
            }
          },
          "_count": {
            "type": "object",
            "properties": {
              "likes": {
                "type": "integer",
                "example": 42
              },
              "comments": {
                "type": "integer",
                "example": 7
              }
            }
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "tags",
          "createdAt",
          "updatedAt",
          "authorId"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmas2xbuy0000tf8cp8h37lgl"
          },
          "name": {
            "type": "string",
            "example": "Jane Developer"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane@example.com"
          },
          "username": {
            "type": "string",
            "example": "janedev"
          },
          "bio": {
            "type": "string",
            "nullable": true,
            "example": "Full-stack developer passionate about TypeScript"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/avatar.jpg"
          },
          "location": {
            "type": "string",
            "nullable": true,
            "example": "San Francisco, CA"
          },
          "website": {
            "type": "string",
            "nullable": true,
            "example": "https://janedev.com"
          },
          "githubUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://github.com/janedev"
          },
          "linkedinUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://linkedin.com/in/janedev"
          },
          "role": {
            "type": "string",
            "example": "USER",
            "default": "USER"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_count": {
            "type": "object",
            "properties": {
              "followers": {
                "type": "integer",
                "example": 125
              },
              "following": {
                "type": "integer",
                "example": 45
              },
              "projects": {
                "type": "integer",
                "example": 12
              }
            }
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "username",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
      "Like": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmas2xbuy0000tf8cp8h37lxx"
          },
          "userId": {
            "type": "string",
            "example": "cmas2xbuy0000tf8cp8h37lgl"
          },
          "projectId": {
            "type": "string",
            "example": "cmas2xbuy0000tf8cp8h37lgl"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": ["id", "userId", "projectId", "createdAt"]
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "data": {
            "type": "object"
          }
        },
        "required": ["success", "message"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "errors": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object"
              }
            ]
          }
        },
        "required": ["success", "message"]
      },
...
    },
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/projects": {
      "get": {
        "summary": "Get paginated list of projects",
        "description": "Retrieves a paginated list of projects with optional filtering, sorting and search",
        "tags": ["Projects"],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search term for filtering projects across title, description and technologies"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Number of projects per page"
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "createdAt",
              "enum": ["createdAt", "title", "updatedAt"]
            },
            "description": "Field to sort by"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "desc",
              "enum": ["asc", "desc"]
            },
            "description": "Sort direction (asc/desc)"
          }
        ],
        "responses": {
          "200": {
            "description": "Projects fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "projects": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Project"
                              }
                            },
                            "meta": {
                              "type": "object",
                              "properties": {
                                "total": {
                                  "type": "integer",
                                  "description": "Total number of projects"
                                },
                                "pages": {
                                  "type": "integer",
                                  "description": "Total number of pages"
                                },
                                "page": {
                                  "type": "integer",
                                  "description": "Current page number"
                                },
                                "limit": {
                                  "type": "integer",
                                  "description": "Number of projects per page"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No projects found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new project",
        "description": "Creates a new project for the authenticated user",
        "tags": ["Projects"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Dev Connect Frontend"
                  },
                  "description": {
                    "type": "string",
                    "example": "Frontend implementation of the Developer Connection platform"
                  },
                  "repoUrl": {
                    "type": "string",
                    "example": "https://github.com/username/dev-connect-frontend"
                  },
                  "liveUrl": {
                    "type": "string",
                    "example": "https://dev-connect-demo.vercel.app"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["React", "TypeScript", "Tailwind CSS"]
                  },
                  "coverImage": {
                    "type": "string",
                    "format": "binary",
                    "description": "Project cover image"
                  }
                },
                "required": ["title", "description"]
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Dev Connect Frontend"
                  },
                  "description": {
                    "type": "string",
                    "example": "Frontend implementation of the Developer Connection platform"
                  },
                  "repoUrl": {
                    "type": "string",
                    "example": "https://github.com/username/dev-connect-frontend"
                  },
                  "liveUrl": {
                    "type": "string",
                    "example": "https://dev-connect-demo.vercel.app"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["React", "TypeScript", "Tailwind CSS"]
                  }
                },
                "required": ["title", "description"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Project"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error or database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Project with same title already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}